move module proxy HTTP client logic to new internal/pkg/modproxyclient package, add "not found" error handling to list
