+++ go/internal/pkg/modproxyclient/error.go
@@ -1,0 +1,7 @@
+package modproxyclient
+
+import (
+	"errors"
+)
+
+var ErrNotFound = errors.New("not found")
